refactor: share detail search rendering
