+++ src/ui/features/browse/mod.rs
@@ -1,4 +1,5 @@
 pub mod cell_detail;
+mod detail_view;
 pub mod explorer;
 pub mod inspector;
 pub mod jsonb_detail;
